backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
python-dotenv
tldextract
confusable-homoglyphs
pytest
supabase
